feat(migration): migrate form vite to Next.js (#8)
