--- src/App.tsx
@@ -1,17 +1,0 @@
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import { LandingPage } from '@/pages/LandingPage';
-import { DashboardPage } from '@/pages/DashboardPage';
-import { ROUTES } from '@/constants/routes';
-
-function App() {
-  return (
-    <BrowserRouter>
-      <Routes>
-        <Route path={ROUTES.LANDING} element={<LandingPage />} />
-        <Route path={ROUTES.DASHBOARD} element={<DashboardPage />} />
-      </Routes>
-    </BrowserRouter>
-  );
-}
-
-export default App;
