+++ src/components/ui/FeatureCard.tsx
@@ -6,9 +6,9 @@ interface FeatureCardProps {
 
 export function FeatureCard({ feature }: FeatureCardProps) {
   return (
-    <div className="bg-white rounded-2xl shadow-sm border border-slate-100 p-8 flex flex-col gap-4 hover:shadow-md transition-shadow duration-200">
-      <span className="text-4xl">{feature.icon}</span>
-      <h3 className="text-lg font-semibold text-slate-800">{feature.title}</h3>
+    <div className="bg-white rounded-2xl shadow-sm border border-slate-100 p-6 sm:p-8 flex flex-col gap-3 sm:gap-4 hover:shadow-md transition-shadow duration-200">
+      <span className="text-3xl sm:text-4xl">{feature.icon}</span>
+      <h3 className="text-base sm:text-lg font-semibold text-slate-800">{feature.title}</h3>
       <p className="text-sm text-slate-500 leading-relaxed">{feature.description}</p>
     </div>
   );
